# Optimize chunkedCopy for sequential writes